+++ src/main/java/zaftnotameni/creatania/mana/flowers/blazunia/BlazuniaBlockStates.java
@@ -1,0 +1,8 @@
+package zaftnotameni.creatania.mana.flowers.blazunia;
+
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+
+public class BlazuniaBlockStates {
+  public static final BooleanProperty HAS_MANA_SOURCE = BooleanProperty.create("has_mana_source");
+  public static final BooleanProperty IS_SUPERHOT = BooleanProperty.create("is_superhot");
+}